refactor: TestFixture 중복 통합
