+++ src/test/java/nextstep/subway/acceptance/path/PathAcceptanceTest.java
@@ -16,7 +16,7 @@ import org.springframework.test.context.ActiveProfiles;
 import java.util.List;
 import java.util.Map;
 
-import static nextstep.subway.acceptance.path.PathAcceptanceTestFixture.*;
+import static nextstep.subway.acceptance.AcceptanceTestFixture.*;
 import static nextstep.subway.utils.AssertUtil.assertResponseCode;
 import static org.assertj.core.api.Assertions.assertThat;
