+++ src/test/java/nextstep/subway/unit/path/PathServiceTest.java
@@ -5,7 +5,7 @@ import nextstep.subway.path.PathService;
 import nextstep.subway.line.PathsResponse;
 import nextstep.subway.station.Station;
 import nextstep.subway.station.StationRepository;
-import nextstep.subway.unit.line.LineTestFixture;
+import nextstep.subway.unit.UnitTestFixture;
 import nextstep.subway.utils.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -15,7 +15,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
-import static nextstep.subway.unit.path.PathTestFixture.*;
+import static nextstep.subway.unit.UnitTestFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("테스트DB 사용한 지하철 경로 서비스 테스트")
@@ -44,10 +44,10 @@ public class PathServiceTest {
     @Test
     void findShortestPathsTest() {
         // given
-        Station 강남역 = stationRepository.save(LineTestFixture.강남역);
-        Station 양재역 = stationRepository.save(LineTestFixture.양재역);
-        Station 교대역 = stationRepository.save(LineTestFixture.교대역);
-        Station 홍대역 = stationRepository.save(LineTestFixture.홍대역);
+        Station 강남역 = stationRepository.save(UnitTestFixture.강남역);
+        Station 양재역 = stationRepository.save(UnitTestFixture.양재역);
+        Station 교대역 = stationRepository.save(UnitTestFixture.교대역);
+        Station 홍대역 = stationRepository.save(UnitTestFixture.홍대역);
 
         lineRepository.save(신분당선(강남역, 양재역));
         lineRepository.save(분당선(양재역, 교대역));
